move Python3 only syntax to its own file, and exclude it in Python2 runs
also, wrap `cause` in a list as required by the APM Server API
