+++ tests/client/exception_tests.py
@@ -291,64 +291,3 @@ def test_message_keyword_truncation(sending_elasticapm_client):
     assert error["log"]["message"] == too_long  # message is not truncated
 
     assert error["log"]["logger_name"] == expected
-
-
-@pytest.mark.skipif(compat.PY2, reason="raise from is Python 3 only")
-def test_chained_cause_exception(elasticapm_client):
-    try:
-        try:
-            1 / 0
-        except ZeroDivisionError as zde:
-            raise ValueError("bla") from zde
-    except ValueError:
-        elasticapm_client.capture_exception()
-    error = elasticapm_client.events[ERROR][0]
-    assert error["exception"]["type"] == "ValueError"
-    assert error["exception"]["cause"]["type"] == "ZeroDivisionError"
-
-
-@pytest.mark.skipif(compat.PY2, reason="raise from is Python 3 only")
-def test_chained_context_exception(elasticapm_client):
-    try:
-        try:
-            1 / 0
-        except ZeroDivisionError:
-            int("zero")
-    except ValueError:
-        elasticapm_client.capture_exception()
-    error = elasticapm_client.events[ERROR][0]
-    assert error["exception"]["type"] == "ValueError"
-    assert error["exception"]["cause"]["type"] == "ZeroDivisionError"
-
-
-@pytest.mark.skipif(compat.PY2, reason="raise from is Python 3 only")
-def test_chained_context_exception_suppressed(elasticapm_client):
-    try:
-        try:
-            1 / 0
-        except ZeroDivisionError:
-            raise ValueError("bla") from None
-    except ValueError:
-        elasticapm_client.capture_exception()
-    error = elasticapm_client.events[ERROR][0]
-    assert error["exception"]["type"] == "ValueError"
-    assert "cause" not in error["exception"]
-
-
-@pytest.mark.skipif(compat.PY2, reason="raise from is Python 3 only")
-def test_chained_context_exception_max(elasticapm_client):
-    with mock.patch("elasticapm.events.EXCEPTION_CHAIN_MAX_DEPTH", 1):
-        try:
-            try:
-                1 / 0
-            except ZeroDivisionError:
-                try:
-                    1 + "a"
-                except TypeError as e:
-                    raise ValueError("bla") from e
-        except ValueError:
-            elasticapm_client.capture_exception()
-        error = elasticapm_client.events[ERROR][0]
-        assert error["exception"]["type"] == "ValueError"
-        assert error["exception"]["cause"]["type"] == "TypeError"
-        assert "cause" not in error["exception"]["cause"]
